fix(patina): split no-duplicate-attr-inheritance tests to satisfy source-length guard
Move the inline #[cfg(test)] module to a sibling no_duplicate_attr_inheritance_tests.rs
(via #[path], matching the script.rs convention) so the rule file is 280 lines (<350).
Formatted with the pinned 1.95.0 toolchain rustfmt (1.9.0-stable, as CI uses).
